Refactor, add tests, and local taskfiles for linting & testing.
