+++ pyproject.toml
@@ -1,0 +1,9 @@
+[tool.pytest.ini_options]
+minversion = "6.0"
+addopts = "--strict-markers"
+testpaths = [
+  "src/tests"
+]
+
+[tool.ruff]
+line-length = 120
