+++ src/collectors/__init__.py
@@ -1,0 +1,5 @@
+from .items import JellyfinItems
+from .sessions import JellyfinSessions
+from .users import JellyfinUsers
+
+__all__ = ["JellyfinItems", "JellyfinSessions", "JellyfinUsers"]
