+++ src/collectors/sessions.py
@@ -1,43 +1,16 @@
-import os
-import logging
-import requests
-import threading
-import time
+from prometheus_client.core import GaugeMetricFamily
 import traceback
-from datetime import datetime
-
-from prometheus_client import start_http_server
-from prometheus_client.core import GaugeMetricFamily, REGISTRY
-
-PORT = int(os.getenv("JELLYFIN_EXPORTER_PORT", 8080))
-URL = os.getenv("JELLYFIN_EXPORTER_URL", "localhost:9090")
-API_KEY = os.getenv("JELLYFIN_EXPORTER_API_KEY", "")
-INSTANCE = os.getenv("JELLYFIN_EXPORTER_INSTANCE_NAME", "jellyfin")
-
-logging.basicConfig(
-    format="%(asctime)s %(levelname)s %(message)s",
-    level=logging.INFO,
-    datefmt="%Y-%m-%d %H:%M:%S",
-)
-
-logging.info("Starting jellyfin_exporter for '%s' on port: %d", str(URL), PORT)
-logging.debug("Starting with api key '%s'", API_KEY)
-
-
-def request_api(action):
-    url = "{}{}?api_key={}".format(URL, action, API_KEY)
-    start = time.time()
-    data = requests.get(url).json()
-    elapsed = time.time() - start
-    logging.info("Request to %s returned in %s", url, elapsed)
-    return data
 
 
 class JellyfinSessions(object):
+    def __init__(self, request_api, critical_logger):
+        self.request_api = request_api
+        self.critical_logger = critical_logger
+
     def collect(self):
         endpoint = "/sessions"
         try:
-            data = request_api(endpoint)
+            data = self.request_api(endpoint)
             yield self.count(data)
             yield self.info(data)
             yield self.bitrate(data)
@@ -49,15 +22,14 @@ class JellyfinSessions(object):
             yield self.transcode_bitrate(data)
             yield self.transcode_completion(data)
         except Exception as ex:
-            logging.critical(f"Error collecting {endpoint}: %s", ex)
-            logging.critical("Traceback: %s", traceback.format_exc())
+            self.critical_logger(f"Error collecting {endpoint}: %s", ex)
+            self.critical_logger("Traceback: %s", traceback.format_exc())
 
     def transcode_info(self, data):
         gauge = GaugeMetricFamily(
             "jellyfin_session_transcode_info",
             "Info about transcoding in the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
                 "audio_codec",
@@ -75,7 +47,6 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                     str(session["TranscodingInfo"].get("AudioCodec", "Unknown")),
@@ -96,7 +67,6 @@ class JellyfinSessions(object):
             "jellyfin_session_transcode_bitrate",
             "The transcoding bitrate of the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
             ],
@@ -106,7 +76,6 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                 ],
@@ -119,7 +88,6 @@ class JellyfinSessions(object):
             "jellyfin_session_transcode_completion",
             "The transcoding completion percentage of the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
             ],
@@ -129,7 +97,6 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                 ],
@@ -142,7 +109,6 @@ class JellyfinSessions(object):
             "jellyfin_session_bitrate",
             "The bitrate of the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
             ],
@@ -153,12 +119,9 @@ class JellyfinSessions(object):
             if "MediaStreams" not in session["NowPlayingItem"]:
                 continue
             streams = session["NowPlayingItem"]["MediaStreams"]
-            bitrate = sum(
-                item.get("BitRate", 0) for item in streams if "BitRate" in item
-            )
+            bitrate = sum(item.get("BitRate", 0) for item in streams if "BitRate" in item)
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                 ],
@@ -171,7 +134,6 @@ class JellyfinSessions(object):
             "jellyfin_session_now_playing_info",
             "Information about the item currently playing on the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
                 "item_id",
@@ -186,7 +148,6 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                     str(session["NowPlayingItem"].get("Id", "Unknown")),
@@ -204,7 +165,6 @@ class JellyfinSessions(object):
             "jellyfin_session_now_playing_run_time",
             "Run time of the item currently playing on the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
                 "item_id",
@@ -215,7 +175,6 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                     str(session["NowPlayingItem"].get("Id", "Unknown")),
@@ -228,21 +187,19 @@ class JellyfinSessions(object):
         gauge = GaugeMetricFamily(
             "jellyfin_sessions",
             "Number of jellyfin sessions",
-            labels=["instance"],
         )
-        gauge.add_metric([INSTANCE], len(data))
+        gauge.add_metric([], len(data))
         return gauge
 
     def info(self, data):
         gauge = GaugeMetricFamily(
             "jellyfin_session_info",
             "Info about current jellyfin sessions.",
-            labels=["instance", "session_id", "user", "client", "device"],
+            labels=["session_id", "user", "client", "device"],
         )
         for session in data:
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                     session["Client"],
@@ -257,7 +214,6 @@ class JellyfinSessions(object):
             "jellyfin_session_play_state_info",
             "Information about the play state of the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
                 "paused",
@@ -270,7 +226,6 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                     str(session["PlayState"].get("IsPaused", "Unknown")),
@@ -286,7 +241,6 @@ class JellyfinSessions(object):
             "jellyfin_session_play_state_position",
             "Information about the play state of the jellyfin session.",
             labels=[
-                "instance",
                 "session_id",
                 "user",
             ],
@@ -296,204 +250,9 @@ class JellyfinSessions(object):
                 continue
             gauge.add_metric(
                 [
-                    INSTANCE,
                     session["Id"],
                     session["UserName"],
                 ],
                 int(session["PlayState"].get("PositionTicks", 0)),
             )
         return gauge
-
-
-class JellyfinUsers(object):
-    def collect(self):
-        endpoint = "/users"
-        try:
-            data = request_api(endpoint)
-            yield self.count(data)
-            yield self.info(data)
-            yield self.invalid_login_attempts(data)
-            yield self.lockout_login_attempts(data)
-            yield self.bitrate_limit(data)
-            yield self.last_login(data)
-        except Exception as ex:
-            logging.critical(f"Error collecting {endpoint}: %s", ex)
-            logging.critical("Traceback: %s", traceback.format_exc())
-
-    def count(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_users",
-            "Number of jellyfin users",
-            labels=["instance"],
-        )
-        gauge.add_metric([INSTANCE], len(data))
-        return gauge
-
-    def info(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_user_info",
-            "Info about jellyfin users.",
-            labels=[
-                "instance",
-                "user_id",
-                "user",
-                "administrator",
-                "hidden",
-                "disabled",
-                "subtitle_language_preference",
-                "subtitle_mode",
-            ],
-        )
-        for user in data:
-            if "Configuration" not in user:
-                continue
-            if "Policy" not in user:
-                continue
-            gauge.add_metric(
-                [
-                    INSTANCE,
-                    user["Id"],
-                    user["Name"],
-                    str(user["Policy"].get("IsAdministrator", "Unknown")),
-                    str(user["Policy"].get("IsHidden", "Unknown")),
-                    str(user["Policy"].get("IsDisabled", "Unknown")),
-                    user["Configuration"].get("SubtitleLanguagePreference", "Unknown"),
-                    user["Configuration"].get("SubtitleMode", "Unknown"),
-                ],
-                1,
-            )
-        return gauge
-
-    def invalid_login_attempts(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_user_invalid_login_attempts",
-            "Number of invalid login attempts for a jellyfin user.",
-            labels=[
-                "instance",
-                "user_id",
-                "user",
-            ],
-        )
-        for user in data:
-            if "Policy" not in user:
-                continue
-            gauge.add_metric(
-                [
-                    INSTANCE,
-                    user["Id"],
-                    user["Name"],
-                ],
-                int(user["Policy"].get("InvalidLoginAttemptCount", 0)),
-            )
-        return gauge
-
-    def lockout_login_attempts(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_user_lockout_login_attempts",
-            "Number of logins attempts before a jellyfin user is locked out.",
-            labels=[
-                "instance",
-                "user_id",
-                "user",
-            ],
-        )
-        for user in data:
-            if "Policy" not in user:
-                continue
-            gauge.add_metric(
-                [
-                    INSTANCE,
-                    user["Id"],
-                    user["Name"],
-                ],
-                int(user["Policy"].get("LoginAttemptsBeforeLockout", 0)),
-            )
-        return gauge
-
-    def bitrate_limit(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_user_bitrate_limit",
-            "The maximum bitrate sessions from a jellyfin user are limited to.",
-            labels=[
-                "instance",
-                "user_id",
-                "user",
-            ],
-        )
-        for user in data:
-            if "Policy" not in user:
-                continue
-            gauge.add_metric(
-                [
-                    INSTANCE,
-                    user["Id"],
-                    user["Name"],
-                ],
-                int(user["Policy"].get("RemoteClientBitrateLimit", 0)),
-            )
-        return gauge
-
-    def _strip_microseconds(self, date_string):
-        """Strip microseconds down to 6 digits if they are longer."""
-        if "." in date_string and "Z" in date_string:
-            before_microseconds, rest = date_string.split(".")
-            microseconds = rest[:-1]  # Remove the 'Z'
-            if len(microseconds) > 6:
-                microseconds = microseconds[:6]
-            return f"{before_microseconds}.{microseconds}Z"
-        return date_string
-
-    def last_login(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_user_last_login",
-            "Last login timestamp of jellyfin users.",
-            labels=[
-                "instance",
-                "user_id",
-                "user",
-            ],
-        )
-        for user in data:
-            if "LastLoginDate" not in user:
-                continue
-            date = self._strip_microseconds(user["LastLoginDate"])
-            gauge.add_metric(
-                [
-                    INSTANCE,
-                    user["Id"],
-                    user["Name"],
-                ],
-                int(datetime.strptime(date, "%Y-%m-%dT%H:%M:%S.%fZ").timestamp()),
-            )
-        return gauge
-
-
-class JellyfinItems(object):
-    def collect(self):
-        endpoint = "/Items/Counts"
-        try:
-            data = request_api(endpoint)
-            yield self.count(data)
-        except Exception as ex:
-            logging.critical(f"Error collecting {endpoint}: %s", ex)
-            logging.critical("Traceback: %s", traceback.format_exc())
-
-    def count(self, data):
-        gauge = GaugeMetricFamily(
-            "jellyfin_items",
-            "Number of jellyfin items.",
-            labels=["instance", "type"],
-        )
-        for item, value in data.items():
-            gauge.add_metric([INSTANCE, item], int(value))
-
-        return gauge
-
-
-REGISTRY.register(JellyfinSessions())
-REGISTRY.register(JellyfinUsers())
-REGISTRY.register(JellyfinItems())
-start_http_server(PORT)
-
-e = threading.Event()
-e.wait()
