+++ src/tests/items/api_1.json
@@ -1,0 +1,14 @@
+{
+    "MovieCount": 325,
+    "SeriesCount": 41,
+    "EpisodeCount": 3517,
+    "ArtistCount": 0,
+    "ProgramCount": 0,
+    "TrailerCount": 0,
+    "SongCount": 0,
+    "AlbumCount": 0,
+    "MusicVideoCount": 0,
+    "BoxSetCount": 56,
+    "BookCount": 0,
+    "ItemCount": 0
+}
